Backend/requirements.txt
flask
flask-cors
pypdf2
python-docx